10/02/23

+++ src/main/java/com/example/appwebbellac/model/Diplome.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class Diplome {
-    private long IdDiplome;
+    private Integer IdDiplome;
 
     private String NOM;
 }
